+++ src/main/java/operations/patterns/NullPattern.java
@@ -1,4 +1,4 @@
-package operations.coding;
+package operations.patterns;
 
 import table.Table;
