+++ src/main/java/operations/patterns/Pattern.java
@@ -1,4 +1,4 @@
-package operations.coding;
+package operations.patterns;
 
 import table.Table;
