+++ src/test/java/operations/coding/SingleOccurrencePatternTest.java
@@ -6,6 +6,8 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.ArrayList;
 
+import operations.patterns.SingleOccurrencePattern;
+
 import org.junit.Before;
 import org.junit.Test;
